refactor: moved IPFS upload function into separate library.
