+++ src/views/Dashboard/Hospital/Account/index.vue
@@ -232,8 +232,8 @@ import { mapGetters } from 'vuex'
 import { updateHospital } from '@/lib/polkadotProvider/command/hospitals'
 import { createCertification, updateCertification, deleteCertification } from '@/lib/polkadotProvider/command/hospitals/certifications'
 import countryData from "@/assets/json/country.json"
-import ipfsWorker from "@/web-workers/ipfs-worker"
 import cityData from "@/assets/json/city.json"
+import { upload } from "@/lib/ipfs/upload"
 import Dialog from '@/components/Dialog'
 import Button from '@/components/Button'
 
@@ -359,7 +359,7 @@ export default {
         const context = this
         fr.addEventListener('load', async () => {
           // Upload
-          const uploaded = await context.upload({
+          const uploaded = await upload({
             fileChunk: fr.result,
             fileType: file.type,
             fileName: file.name,
@@ -373,42 +373,6 @@ export default {
         this.files = []
         this.image = ''
       }
-    },
-    upload({ fileChunk, fileName, fileType }) {
-      const chunkSize = 10 * 1024 * 1024 // 10 MB
-      let offset = 0
-      const blob = new Blob([ fileChunk ], { type: fileType })
-
-      return new Promise((resolve, reject) => {
-        try {
-          const fileSize = blob.size
-          do {
-            let chunk = blob.slice(offset, chunkSize + offset);
-            ipfsWorker.workerUpload.postMessage({
-              seed: chunk.seed, file: blob
-            })
-            offset += chunkSize
-          } while((chunkSize + offset) < fileSize)
-          
-          let uploadSize = 0
-          const uploadedResultChunks = []
-          ipfsWorker.workerUpload.onmessage = async event => {
-            uploadedResultChunks.push(event.data)
-            uploadSize += event.data.data.size
-              
-            if (uploadSize >= fileSize) {
-              resolve({
-                fileName: fileName,
-                fileType: fileType,
-                ipfsPath: uploadedResultChunks
-              })
-            }
-          }
-
-        } catch (err) {
-          reject(new Error(err.message))
-        }
-      })
     },
     closeCertificationDialog() {
       this.certId = ""
