+++ src/views/Dashboard/Hospital/Registration/index.vue
@@ -110,8 +110,8 @@ import { registerHospital } from "@/lib/polkadotProvider/command/hospitals"
 import { setEthAddress } from "@/lib/polkadotProvider/command/userProfile"
 import { getWalletAddress } from "@/lib/metamask/wallet"
 import countryData from "@/assets/json/country.json"
-import ipfsWorker from "@/web-workers/ipfs-worker"
 import cityData from "@/assets/json/city.json"
+import { upload } from "@/lib/ipfs/upload"
 
 export default {
   name: "HospitalRegistration",
@@ -220,7 +220,7 @@ export default {
         const context = this
         fr.addEventListener("load", async () => {
           // Upload
-          const uploaded = await context.upload({
+          const uploaded = await upload({
             fileChunk: fr.result,
             fileType: file.type,
             fileName: file.name,
@@ -234,42 +234,6 @@ export default {
         this.files = []
         this.image = ""
       }
-    },
-    upload({ fileChunk, fileName, fileType }) {
-      const chunkSize = 10 * 1024 * 1024 // 10 MB
-      let offset = 0
-      const blob = new Blob([ fileChunk ], { type: fileType })
-
-      return new Promise((resolve, reject) => {
-        try {
-          const fileSize = blob.size
-          do {
-            let chunk = blob.slice(offset, chunkSize + offset);
-            ipfsWorker.workerUpload.postMessage({
-              seed: chunk.seed, file: blob
-            })
-            offset += chunkSize
-          } while((chunkSize + offset) < fileSize)
-          
-          let uploadSize = 0
-          const uploadedResultChunks = []
-          ipfsWorker.workerUpload.onmessage = async event => {
-            uploadedResultChunks.push(event.data)
-            uploadSize += event.data.data.size
-              
-            if (uploadSize >= fileSize) {
-              resolve({
-                fileName: fileName,
-                fileType: fileType,
-                ipfsPath: uploadedResultChunks
-              })
-            }
-          }
-
-        } catch (err) {
-          reject(new Error(err.message))
-        }
-      })
     },
   }
 }
