+++ src/views/Dashboard/Lab/Services/Add.vue
@@ -71,7 +71,7 @@
 
 <script>
 import { mapGetters } from 'vuex'
-import ipfsWorker from '@/web-workers/ipfs-worker'
+import { upload } from "@/lib/ipfs/upload"
 import { createService } from '@/lib/polkadotProvider/command/services'
 
 export default {
@@ -128,7 +128,7 @@ export default {
         const context = this
         fr.addEventListener('load', async () => {
           // Upload
-          const uploaded = await context.upload({
+          const uploaded = await upload({
             fileChunk: fr.result,
             fileType: file.type,
             fileName: file.name,
@@ -142,42 +142,6 @@ export default {
         this.files = []
         this.imageUrl = ''
       }
-    },
-    upload({ fileChunk, fileName, fileType }) {
-      const chunkSize = 10 * 1024 * 1024 // 10 MB
-      let offset = 0
-      const blob = new Blob([ fileChunk ], { type: fileType })
-
-      return new Promise((resolve, reject) => {
-        try {
-          const fileSize = blob.size
-          do {
-            let chunk = blob.slice(offset, chunkSize + offset);
-            ipfsWorker.workerUpload.postMessage({
-              seed: chunk.seed, file: blob
-            })
-            offset += chunkSize
-          } while((chunkSize + offset) < fileSize)
-          
-          let uploadSize = 0
-          const uploadedResultChunks = []
-          ipfsWorker.workerUpload.onmessage = async event => {
-            uploadedResultChunks.push(event.data)
-            uploadSize += event.data.data.size
-              
-            if (uploadSize >= fileSize) {
-              resolve({
-                fileName: fileName,
-                fileType: fileType,
-                ipfsPath: uploadedResultChunks
-              })
-            }
-          }
-
-        } catch (err) {
-          reject(new Error(err.message))
-        }
-      })
     },
   },
 }
